backend/app_recipe/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.1
pandas==2.1.3
numpy==1.26.2
python-multipart==0.0.6
starlette==0.27.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
uvloop==0.19.0 